+++ advtraj/utils/grid.py
@@ -44,6 +44,8 @@ def wrap_posn(x, x_min, x_max):
 
 
 def find_coord_grid_spacing(da_coord, show_warnings=True):
+    grid_tol = 0.001
+    
     v_name = f"d{da_coord.name}"
     if v_name in da_coord.attrs:
         return da_coord.attrs[v_name]
@@ -58,10 +60,10 @@ def find_coord_grid_spacing(da_coord, show_warnings=True):
 
     dx_all = np.diff(da_coord.values)
 
-    if np.min(dx_all) != np.max(dx_all):
+    if (np.max(dx_all) - np.min(dx_all)) / np.mean(dx_all) > grid_tol :
         raise Exception("Non-uniform grid")
 
-    return dx_all[0]
+    return np.mean(dx_all)
 
 
 def find_grid_spacing(ds_grid, coords=("x", "y", "z")):